+++ lib/grape_token_auth/responses/base.rb
@@ -1,0 +1,11 @@
+module GrapeTokenAuth
+  module Responses
+    class Base
+      attr_reader :status_code
+
+      def initialize(status_code = 200)
+        @status_code = status_code
+      end
+    end
+  end
+end