+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current
 
+- Add ability to override resource responses
 - Add frozen string comments for ruby 2.3
 - Fix several bugs around rails sign out and session
 - Fix insensitive_keys bug with devise compatability